requirements.txt

mysql-connector-python
hvac
